remove Layer.java class and remove 'policy' property of a Rollout
