--- core-api/src/main/java/com/optimizely/ab/config/Layer.java
@@ -1,83 +1,0 @@
-/**
- *
- *    Copyright 2017, Optimizely and contributors
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-package com.optimizely.ab.config;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.util.List;
-
-import javax.annotation.concurrent.Immutable;
-
-/**
- * Represents a Optimizely Layer configuration
- *
- * @see <a href="http://developers.optimizely.com/server/reference/index.html#json">Project JSON</a>
- */
-@Immutable
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Layer implements IdMapped {
-
-    protected final String id;
-    protected final String policy;
-    protected final List<Experiment> experiments;
-    public enum LayerPolicy {
-        ROLLOUT ("rollout"),
-        SINGLE_EXPERIMENT ("single_experiment");
-
-        private final String layerPolicy;
-
-        LayerPolicy(String layerPolicy) {
-            this.layerPolicy = layerPolicy;
-        }
-
-        public String toString() {
-            return layerPolicy;
-        }
-    }
-
-    @JsonCreator
-    public Layer(@JsonProperty("id") String id,
-                 @JsonProperty("policy") String policy,
-                 @JsonProperty("experiments") List<Experiment> experiments) {
-        this.id = id;
-        this.policy = policy;
-        this.experiments = experiments;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public String getPolicy() {
-        return policy;
-    }
-
-    public List<Experiment> getExperiments() {
-        return experiments;
-    }
-
-    @Override
-    public String toString() {
-        return "Layer{" +
-                "id='" + id + '\'' +
-                ", policy='" + policy + '\'' +
-                ", experiments=" + experiments +
-                '}';
-    }
-}
